+++ src/pages/attestering/attesterSøknadsbehandling/attesterSøknadsbehandling-nb.ts
